+++ packages/web-platform/web-explorer/index.ts
@@ -1,8 +1,5 @@
-import '@lynx-js/web-core';
-import type { LynxView } from '@lynx-js/web-core';
-import '@lynx-js/web-core/index.css';
-import '@lynx-js/web-elements/index.css';
-import '@lynx-js/web-elements/all';
+import '@lynx-js/web-core/client';
+import type { LynxViewElement as LynxView } from '@lynx-js/web-core/client';
 
 const video = document.getElementById('qr-scanner') as HTMLVideoElement;
 let lynxView = document.getElementById('lynx-view') as LynxView;