+++ packages/webpack/cache-events-webpack-plugin/test/cases/cache-events/chunk-splitting/index.js
@@ -5,10 +5,10 @@ import { add } from './lib-common.js';
 it('should have `__webpack_require__.lynx_ce`', () => {
   expect(add(1, 2)).toBe(3);
   expect(__webpack_require__['lynx_ce']).toBeTruthy();
-  expect(__webpack_require__['lynx_ce']['setupList'].length).toBe(2);
+  expect(__webpack_require__['lynx_ce']['setupList'].length).toBe(3);
   expect(
     __webpack_require__['lynx_ce']['setupList'].map((item) => item.name),
-  ).toEqual(['ttMethod', 'performanceEvent']);
+  ).toEqual(['ttMethod', 'performanceEvent', 'globalThis']);
   expect(__webpack_require__['lynx_ce']['loaded']).toBe(true);
   expect(__webpack_require__['lynx_ce']['cachedActions'].length).toBe(
     0,